requirements.txt
pytest==8.2.2
requests==2.32.3
allure-pytest==2.13.5
pyyaml==6.0.2
pytest-html==4.1.1
